+++ src/main/java/com/programmers/voucher/service/VoucherService.java
@@ -1,0 +1,4 @@
+package com.programmers.voucher.service;
+
+public class VoucherService {
+}
