+++ src/main/java/com/programmers/voucher/util/TriFunction.java
@@ -1,0 +1,6 @@
+package com.programmers.voucher.util;
+
+public interface TriFunction<A, B, C, R> {
+
+    R apply(A a, B b, C c);
+}
