+++ src/test/java/com/programmers/voucher/VoucherApplicationTests.java
@@ -1,4 +1,4 @@
-package com.programmers.domain;
+package com.programmers.voucher;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
