+++ src/test/java/com/programmers/voucher/domain/DiscountAmountTest.java
@@ -1,10 +1,8 @@
-package com.programmers.domain.voucher;
+package com.programmers.voucher.domain;
 
 import com.programmers.exception.AmountValueException;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.springframework.beans.factory.annotation.Value;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
